+++ Sources/ViFlow/macOS/ModernKey/AppDelegate.h
@@ -1,0 +1,35 @@
+//
+//  AppDelegate.h
+//  ModernKey
+//
+//  Created by Tuyen on 1/18/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+#import "ViewController.h"
+
+#define OPENKEY_BUNDLE @"com.quangdv09.viflow"
+
+@interface AppDelegate : NSObject <NSApplicationDelegate, NSWindowDelegate>
+
+-(void)onImputMethodChanged:(BOOL)willNotify;
+-(void)onInputMethodSelected;
+
+-(void)askPermission;
+
+-(void)onInputTypeSelectedIndex:(int)index;
+-(void)onCodeTableChanged:(int)index;
+
+-(void)setRunOnStartup:(BOOL)val;
+-(void)loadDefaultConfig;
+
+-(void)setGrayIcon:(BOOL)val;
+
+-(void)onMacroSelected;
+-(void)onQuickConvert;
+-(void)setQuickConvertString;
+
+-(void)showIconOnDock:(BOOL)val;
+@end
+