+++ Sources/ViFlow/macOS/ModernKey/MJAccessibilityUtils.h
@@ -1,0 +1,21 @@
+//
+//  MJAccessibilityUtils.h
+//  OpenKey
+//
+//  Created by Nguyen Tan Thong on 18/9/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+//  Source: https://github.com/Hammerspoon/hammerspoon/blob/master/Hammerspoon/MJAccessibilityUtils.h
+//  License: MIT
+
+#ifndef MJAccessibilityUtils_h
+
+#import <Foundation/Foundation.h>
+
+BOOL MJAccessibilityIsEnabled(void);
+void MJAccessibilityOpenPanel(void);
+
+#define MJAccessibilityUtils_h
+
+
+#endif /* MJAccessibilityUtils_h */