+++ Sources/ViFlow/macOS/ModernKey/MacroViewController.h
@@ -1,0 +1,23 @@
+//
+//  MacroViewController.h
+//  OpenKey
+//
+//  Created by Tuyen on 8/4/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface MacroViewController : NSViewController<NSTableViewDataSource, NSTableViewDelegate, NSTextFieldDelegate>
+@property (weak) IBOutlet NSTableView *tableView;
+@property (weak) IBOutlet NSTextField *macroName;
+@property (weak) IBOutlet NSTextField *macroContent;
+
+@property (weak) IBOutlet NSButton *buttonAdd;
+@property (weak) IBOutlet NSButton *AutoCapsMacro;
+
+@end
+
+NS_ASSUME_NONNULL_END