+++ Sources/ViFlow/macOS/ModernKey/main.m
@@ -1,0 +1,14 @@
+//
+//  main.m
+//  ModernKey
+//
+//  Created by Tuyen on 1/18/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+#import "ViewController.h"
+
+int main(int argc, const char * argv[]) {
+    return NSApplicationMain(argc, argv);
+}