+++ Sources/ViFlow/macOS/ViFlowHelper/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  OpenKeyHelper
+//
+//  Created by Tuyen on 2/1/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface AppDelegate : NSObject <NSApplicationDelegate>
+
+
+@end
+