+++ Sources/ViFlow/macOS/ViFlowHelper/main.m
@@ -1,0 +1,13 @@
+//
+//  main.m
+//  OpenKeyHelper
+//
+//  Created by Tuyen on 2/1/19.
+//  Copyright © 2026 Do Vinh Quang. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+int main(int argc, const char * argv[]) {
+    return NSApplicationMain(argc, argv);
+}